+++ middlewares/04_bodyParserJSON.js
@@ -1,2 +1,0 @@
-logger.log("Setting bodyParser JSON");
-module.exports =  bodyParser.json();