+++ middlewares/10_passportSession.js
@@ -1,7 +1,0 @@
-if (CONFIG.enablePassport) {
-  module.exports = global.passport.session();
-} else {
-  module.exports = function(req, res, next) {
-    next();
-  }
-}