+++ middlewares/11_flashMessages.js
@@ -1,7 +1,0 @@
-if (CONFIG.enableRedis) {
-  module.exports =  global.flash({ locals: 'flash' });
-} else {
-  module.exports = function(req, res, next) {
-    next();
-  }
-}