+++ middlewares/bodyParserJSON.js
@@ -1,0 +1,2 @@
+logger.log("Setting bodyParser JSON");
+module.exports =  require('body-parser').json();