+++ middlewares/flashMessages.js
@@ -1,0 +1,7 @@
+if (CONFIG.session !== false) {
+  module.exports =  require('req-flash')({ locals: 'flash' });
+} else {
+  module.exports = function(req, res, next) {
+    next();
+  }
+}